+++ src/pongstarBase.cpp
@@ -70,7 +70,7 @@ void PongstarBase::initializeEntities() {
 	// For pickups testing
 	// std::vector<effectNS::EFFECT_TYPE> ev = { effectNS::SHIELD, effectNS::MULTIPLY };
 	// pickupManager->testPickup(effectNS::MAGNET);
-	// pickupManager->createPickup(effectNS::MAGNET);
+	// pickupManager->createPickup(effectNS::INVERT);
 	// pickupManager->massSpawnPickups(ev);
 	// pickupManager->massSpawnContrastPickups();
 }
@@ -112,7 +112,7 @@ void PongstarBase::update(float frameTime) {
 	
 	if (gameStarted) {
 		steady_clock::time_point presentTime = steady_clock::now();
-		elapsedTime = duration_cast<milliseconds>(presentTime - startTime).count();
+		elapsedTime = (int)duration_cast<milliseconds>(presentTime - startTime).count();
 	}
 
 	for (auto& x : *pickupDelayTimers) {
